change todo migration and create with all users
