+++ app/Models/Todo.php
@@ -15,9 +15,14 @@ class Todo extends Model
 
 
     protected $guarded = ['id'];
+    protected $dates = ['date'];
 
     public function user(){
         return $this->belongsTo(User::class);
+    }
+
+    public function replayTodos(){
+        return $this->hasMany(Todo::class, 'todo_id');
     }
