+++ public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/js/app.js": "/js/app.js?id=ddc9ab5e33aa436ea016150ddabd21c3",
+    "/js/app.js": "/js/app.js?id=93a35d6510906f63fa599defad009f7b",
     "/css/app.css": "/css/app.css?id=8a0ad0de605665aed0683cb06ef1b5a5"
 }
